feat: start multiple files with entrypoint.sh allowed app to starting jobs
