+++ Dockerfile
@@ -14,7 +14,9 @@ COPY .env /app/.env
 COPY ormconfig-docker.json /app/ormconfig.json
 COPY tsconfig.json /app/
 COPY heroku.yml /app/
+COPY entrypoint.sh /app/
 COPY ./src /app/src
 
 RUN npm run tsc
-CMD ["npm", "run", "build"]
+# CMD ["npm", "run", "build"]
+CMD ["/bin/bash", "/app/entrypoint.sh"]
