+++ entrypoint.sh
@@ -1,0 +1,3 @@
+#!/app/bin/env bash 
+node /app/build/index.js &
+node /app/build/jobs.js
